combine toggleCheckbox with setupApp
