+++ superset/assets/src/setup/setupApp.js
@@ -2,7 +2,31 @@
 import $ from 'jquery';
 import { SupersetClient } from '@superset-ui/core';
 import setupClient from './setupClient';
-import toggleCheckbox from '../utils/toggleCheckbox';
+import getClientErrorObject from '../utils/getClientErrorObject';
+
+function showApiMessage(resp) {
+  const template =
+    '<div class="alert"> ' +
+    '<button type="button" class="close" ' +
+    'data-dismiss="alert">\xD7</button> </div>';
+  const severity = resp.severity || 'info';
+  $(template).addClass('alert-' + severity)
+    .append(resp.message)
+    .appendTo($('#alert-container'));
+}
+
+function toggleCheckbox(apiUrlPrefix, selector) {
+  SupersetClient.get({ endpoint: apiUrlPrefix + $(selector)[0].checked })
+    .then(() => {})
+    .catch(response =>
+      getClientErrorObject(response)
+        .then((parsedResp) => {
+          if (parsedResp && parsedResp.message) {
+            showApiMessage(parsedResp);
+          }
+        }),
+      );
+}
 
 export default function setupApp() {
   setupClient();
